move default expansion + filtering to lucuma.odb.sequence
